Add files via upload
The data reflects some interesting outputs, the maximum number of hospitalized cases in the UK Gibraltar area was 5707 and the number of hospitalized cases was 4907. the recovered cases were also 4670 which was a good insight into how many people were able to recover 
a maximum number of death recorded was 97. the data for covid cases specific to Gibraltar was overall a part of the overall covid cases in the uk region.